Refactor code in autohint.py and otffont.py to move font-format specific calls and properties into otffont.py, per Khaled's review.
